Guardamos la ruta de la imagen de perfil por defecto en la base de datos, para asegurarnos de que todas las imágenes carguen correctamente. Modificamos la sentencia de eliminación, para no eliminar la imagen por defecto, en caso de que el usuario a eliminar la esté utilizando.

+++ src/models/delete-data.php
@@ -21,13 +21,15 @@
         if ($row['count'] == 1) {
             // Obtener la ruta de la imagen de perfil
             $file_path = $row['fotografia'];
+            // Ruta de la imagen por defecto
+            $default_image_path = "../../public/img/profile-default.svg";
 
             // Eliminar el perfil del usuario
             $sql = "DELETE FROM usuarios WHERE usuario = '$user' AND contrasenia = '$cryptPass'";
             $result = $conn->query($sql);
 
             // Eliminamos la imagen del usuario del servidor en caso de que exista
-            if (file_exists($file_path)) {
+            if (file_exists($file_path) && $file_path !== $default_image_path) {
                 unlink($file_path);
             }
